Deduplicate device_info plumbing across sensor/number/button platforms
Extract CubicSecureEntityMixin (_attr_attribution, _attr_has_entity_name,
device_info) in __init__.py, alongside cubic_secure_device_info, and have
AbstractLkCubicSensor (sensor.py), LKPauseLeakDetectionDurationNumber
(number.py), and _LKCubicSecureButton (button.py) inherit from it instead
of each redefining the same three members independently.

Standardizes sensor.py's device identity attribute name from _id to
_device_identity to match the other two platforms, so the mixin has one
consistent attribute to read across all three.

Pure refactor, no behavior change - existing per-platform
belongs_to_the_cubic_secure_device tests already pinned device_info's
effective behavior and stay green throughout.
